Fix running tests outside of a jar and remove forgotten assert
Eclipse runs tests with classes not packed in a jar. It is needed to add
"file:" to the repo resource path or it fails in
SourceManagerImpl#createSourceURI on line 103 - a MalformedURLException
is risen because of no protocol specified.

The assert was probably forgotten in the code, because it would limit
the deps.txt size in other tests than simple-dep to 2 entries, which
doesn't make much sense.