Make the code easy on the eye
The continue statement caused that the code in else block was executed
under the same cisrcumstances as the code after the if-else, but it was
not clear at first sight.

So moving the code after the if-else into the else block and dropping
the continue statement seems like a reasonable step to me.