Make test jar files empty
I think there is no need to have any content in the jars, just their
presence is enough. If we keep the same content as in the pom files,
then it would be confusing if we change the pom file and the contents
would be different or we would have to keep them in sync.