Do not close the graph connection too early
It doesn't make much sense to close the graph (connection) in the
extractor because it was not opened there. When we closed the connection
at that point it was not available when collected relationships were
processed below and it led to InvalidRefException.

The extractor is used only in resolveAndExtractMultiGraph which opens and
closes all graphs used, so it is counter-productive to try to help it.