Group paths by the target, not by the starting artifact
It was like that in former versions and this makes the reports in the
Maven Repository Builder buggy. The paths are used there (and possibly
anywhere else) to identify which relationships paths lead to a desired
artifact. Therefore it is handy to get them groupped by the targets,
not by roots.