Add missing equals/hashcode methods in filters
Those missing methods were indirectly causing OutOfMemoryError in paths
method. The filters were never considered to be in the filters of actual
OrFilter, so the same filter was added many times and in the end it
failed even when I started Cartographer with Xmx 32 GB.

Also added caching of scope in NeoDependencyRelationship to not spit the
resolved scope message in the log with every call (called from multiple
places for a single relationship).